feat: :sparkles: Update google_service_account_iam_binding members
This commit updates the `google_service_account_iam_binding` resource to use the `var.repository` list to dynamically generate the `members` attribute. Each repository in the list is now added as a member with the format `"principalSet://iam.googleapis.com/${var.pool_name}/attribute.repository/${repo}"`. This change allows for more flexibility in managing the members of the service account IAM binding.